:truck: [Rename]: Member 클래스 member 패키지로 이동
