+++ src/main/java/com/goormthonuniv/ownearth/domain/mapping/MemberItem.java
@@ -3,8 +3,8 @@ package com.goormthonuniv.ownearth.domain.mapping;
 import jakarta.persistence.*;
 
 import com.goormthonuniv.ownearth.domain.Item;
-import com.goormthonuniv.ownearth.domain.Member;
 import com.goormthonuniv.ownearth.domain.common.BaseEntity;
+import com.goormthonuniv.ownearth.domain.member.Member;
 
 import lombok.*;
